code organization and add refresh to scroll
