+++ analysis_options.yaml
@@ -27,11 +27,3 @@ linter:
 
 # Additional information about this file can be found at
 # https://dart.dev/guides/language/analysis-options
-
-
-analyzer:
-  exclude:
-    - "**/*.g.dart"
-    - "**/*.freezed.dart"
-  errors:
-    invalid_annotation_target: ignore
