--- lib/models/constant.dart
@@ -1,4 +1,0 @@
-class Constant {
-  static const issueStateOpen = 'open';
-  static const issueStateClose = 'closed';
-}
