--- lib/pages/widget/loading.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Loading extends StatelessWidget {
-  const Loading({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const CircularProgressIndicator();
-  }
-}
