+++ lib/pages/widget/sort_button.dart
@@ -11,46 +11,41 @@ class SortButton extends StatelessWidget {
       alignment: Alignment.centerRight,
       child: IconButton(
         onPressed: () {
-          showDialog<void>(
-              builder: (context) => Padding(
-                    padding: const EdgeInsets.all(10),
-                    child: AlertDialog(
-                      content: SingleChildScrollView(
-                        child: ListBody(
-                          children: const [
-                            Text('1年以上更新のないIssueを除外する'),
-                            Text('closed状態のIssueを除外する'),
-                            Text('作成日時の新しい順'),
-                            Text('最終更新日時の古い順'),
-                            Text('コメント数の多い順'),
-                          ],
-                        ),
+          AlertDialog(
+            content: SingleChildScrollView(
+              child: ListBody(
+                children: const [
+                  Text('1年以上更新のないIssueを除外する'),
+                  Text('closed状態のIssueを除外する'),
+                  Text('作成日時の新しい順'),
+                  Text('最終更新日時の古い順'),
+                  Text('コメント数の多い順'),
+                ],
+              ),
+            ),
+            actions: <Widget>[
+              Center(
+                child: SizedBox(
+                  width: double.infinity,
+                  child: Padding(
+                    padding: const EdgeInsets.symmetric(horizontal: 30.0),
+                    child: ElevatedButton(
+                      child: const Text(
+                        '適用する',
+                        style: TextStyle(fontWeight: FontWeight.bold),
                       ),
-                      actions: <Widget>[
-                        Center(
-                          child: SizedBox(
-                            width: double.infinity,
-                            child: Padding(
-                              padding: const EdgeInsets.symmetric(horizontal: 30.0),
-                              child: ElevatedButton(
-                                child: const Text(
-                                  '適用する',
-                                  style: TextStyle(fontWeight: FontWeight.bold),
-                                ),
-                                style: ElevatedButton.styleFrom(
-                                  primary: Colors.grey[400],
-                                  onPrimary: Colors.black,
-                                  shape: const StadiumBorder(),
-                                ),
-                                onPressed: () => Navigator.pop(context),
-                              ),
-                            ),
-                          ),
-                        ),
-                      ],
+                      style: ElevatedButton.styleFrom(
+                        primary: Colors.grey[400],
+                        onPrimary: Colors.black,
+                        shape: const StadiumBorder(),
+                      ),
+                      onPressed: () => Navigator.pop(context),
                     ),
                   ),
-              context: context);
+                ),
+              ),
+            ],
+          );
         },
         icon: const Icon(Icons.filter_list),
       ),
