+++ pubspec.yaml
@@ -36,12 +36,6 @@ dependencies:
   cupertino_icons: ^1.0.2
   http: ^0.13.4
   intl: ^0.17.0
-  url_launcher: ^6.0.20
-  freezed_annotation:
-  json_serializable:
-  flutter_state_notifier: ^0.7.1
-  fluttertoast: ^8.0.9
-  logger: ^1.1.0
 
 dev_dependencies:
   flutter_test:
@@ -53,8 +47,6 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^1.0.0
-  build_runner:
-  freezed:
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
